Fix malfunctioning sigmoid kernel in 16x8 quantization
This CL fixes overflowing sigmoid kernel in 16x8 quantization, which is `reference_integer_op::Logistic()`. Before this CL, it has uint8_t uh variable, which makes overflowing and periodic outputs on sigmoid values. As `reference_integer_op::Tanh()` did, this CL defines uh as uint32_t and saturates it to the maximum 07FFF << 10 if uh overflows over sigmoid_table_uint8[255] (common.h). The maximum value is defined from the reverse engineering of the negation transformation at the last of Logistic(), which is similar to that of Tanh(). In details, since Tanh() adds 1<<7 and right shifts 8, maximum is defined as 0xFFFF << 8. Likewise, since Logistic() adds 1<<9 and right shifts 10, maximum is defined as 0x7FFF << 10. (Here, the reason why it has less 1 bit is Logistic() should be half the value due to the definition, tanh(x)=2*sigmoid(2*x) - 1).

PiperOrigin-RevId: 338420846
Change-Id: I616f8f574a1436d35d63ce7fbdac7ce1ce749f62