SidecarEvaluator: Create a Keras SidecarEvaluator API for a dedicated evaluator task.
This is targeting CTL users running an evaluator job, as opposed to alternating train/evaluate: 1) reading checkpoints as one becomes available, 2) evaluating the model with the dataset set aside for evaluation, 3) logging the evaluated metric, optionally generating summary file for TensorBoard, and optionally exporting the model considered the best so far for serving (TODO). It is possible for users to create such custom loop for evaluation, but this utility may be useful for standard, continuous evaluation that takes care of some details. This is also for feature parity with estimator.

PiperOrigin-RevId: 339623940
Change-Id: I057fd255d46543003217ebbdc339155717c06044