[XLA] Improve memory space assmt compile time by reducing # of gte's and tuples temporarily created
The memory space assignment pass first determines which buffers will get
alternate memory allocations, then modifies the graph (process) to swap the old
buffers with new buffers, and then simplifies the graph by removing dead code
and tuples that gte's all the elements from a previous tuple. Whenever the
process step needed to refer to a shape within a tuple shape, it would create a
get-tuple-element instruction, since there was a simplification step afterwards.
However, in some cases, this created a lot of instructions (in one example
consisting of ~8K HLOs, there were ~12M get-tuple-element and tuple
instructions, which were later simplified back to ~8K). This approach slowed
compilation down, and occasionally caused compiler OOM.

To fix this, this CL tries to reuse existing get-tuple-elements and tuple
instructions in the process step.

It also includes a test case for b/172243149, which was due to my previous
attempt at improving the compilation time and compiler memory usage.

PiperOrigin-RevId: 340568495
Change-Id: If28948e2d8f108d0b804a125a0795980319c4898