Fix issue preventing debug mode from working.
The issue was some constants were declared but not defined. When optimizations were enabled, the constants would be inlined which is why this wasn't causing issues in opt mode. @timshen91 helped me debug this issue.

This was reproducable by running the following command to build the pip package, then installing the pip package and running `import tensorflow as tf`:

    bazel build --config=dbg  //tensorflow/tools/pip_package:build_pip_package

The error was:

    ImportError: /home/reedwm/venvs/tf_dbg/lib/python3.6/site-packages/tensorflow/python/_pywrap_tensorflow_internal.so: undefined symbol: _ZN10tensorflow4data12experimental19SnapshotDatasetV2Op5kHashE

PiperOrigin-RevId: 341209683
Change-Id: If71c6ca529bd87cd1a589422440454548165e813