Update metrics to better support tensors that are not fully defined
If a tensor is not fully defined, broadcast_weights calls
DenseToDenseSetOperation which isn't supported by TPUs. The
DenseToDenseSetOperation is called during an assert before the
weights are broadcasted to check if the values and the weights
have valid dimensions for broadcasting.

TPUs do not support asserts so executing the assert when on
a TPU just takes time without having any beneficial effect.
Instead, on TPUs / GPUs, skip the assert.

PiperOrigin-RevId: 341845786
Change-Id: Ibcacc025df5a35fe1a3ad4357547623832bc7f55