Add aliasing info to serialized module for better memory management on device.
Set `tf.aliasing_output` attribute for serialized module inputs and outputs that are read and written to the same resource. This is passed as `input_output_alias` config to the compiled HLO module and helps it share buffer between inputs and outputs.

`tpu-merge-variabels-with-execute` pass already does this analysis and merges resource reads and writes with TPUExecute op. In this change, we update the serialized module with this alias information as well.

PiperOrigin-RevId: 341878702
Change-Id: I010e6ffec427321b763f75f9e5df9cfe800da2ad