Improve clustering of outside compiled ops to check for circular dependencies.
Outside compilation clusters can be merged together as long as there are no data or side effecting ops interposed between the clusters.  This is checked when adding ops to a cluster by using a DFS to check that there isn't a data/side effect path between the op and the cluster it is to be added to.  This improves on the existing algorithm by correctly recognizing these restrictions and merging more ops in the same cluster when possible.

PiperOrigin-RevId: 342748420
Change-Id: Ic810d1fd6fb6425aa494d3d1f375a0cdaa8c33cb