[TF-TRT] Simplify the parameters for the TF-TRT converters.
Remove the deprecated session_config parameter for the TF1 converter.
Previously, we issued a warning when the value of the parameter is not None.
Change the implementation of the calibrate API to allow soft placement when
running the graph for calibration. This allows OPs that are supported by
TensorRT but do not have an implementation on GPUs to execute on CPUs.

Remove three deprecated fields from the parameter struct for the TF2 converter.
They are: rewriter_config_template, is_dynamic_op and max_batch_size.
Previously, we issued a warning when rewriter_config_template is not None. We
issued an error when is_dynamic_op is not True. We do not use the value of
max_batch_size.

Add two intended to be protected fields, _test_only_disable_non_trt_optimizers
and _test_only_use_implicit_batch, to the TF-TRT converters, to turn off
non-TF-TRT optimizers and switch to explicit batch mode to support TF-TRT
testing.

Modify TF-TRT tests to work with the changes.

PiperOrigin-RevId: 343318190
Change-Id: Ifdbbe8a8cfd827404c7c8e1dd19e041679bd9a96