Update Graph -> TF MLIR importer to always import tf.While output shapes.
output_shapes attribute is reverted back to be a derived attribute and output shapes are populated as result types on import. A new attribute, shape_invariant, is added to tf.While/tf.WhileRegion indicating whether shapes were originally present on the While op. This attribute will be used to update the shape inference for tf.While/tf.WhileRegion, where it is possible to have different result shapes and operand shapes cannot simply be propagated as result shapes.

PiperOrigin-RevId: 343546219
Change-Id: Ib7e832f11edabd59d0f3b45cb0d31c6b6187e706