Infer quantization parameters from the constant and activation ops at the same time
Inferring quantization parameters from the constant and activation ops might
help to quantize some of the layers, but we should use a same flag to turn on
and off this feature. With this flag being turned off, the quantizer only uses
the quantization parameters from the quantization emulation ops. Since the
calibrator doesn't record the ranges of these tensors, this feature should be
turned on in the post-training quantization workflow.

PiperOrigin-RevId: 343930843
Change-Id: I804a339c9d03b2fad9f21c1b8f8c40df14beac79