Add support for returning CompositeTensors from vectorized functions.
This applies semantics similar to those used by `tf.map_fn`:
  - Within the loop body, any returned CompositeTensors are unpacked as tensors.
  - Outside the loop body, the now-stacked Tensors are reassembled into the original composite type.
This works only for CompositeTensors that implement `BatchableTypeSpec`.

This CL doesn't attempt to address passing CompositeTensors as *inputs* to vectorized_map, as demonstrated by edloper@ in cl/337311849, but I believe it is compatible with the approach shown there.

The dummy `Particle` class in the tests is a toy model of a TFP Distribution or Bijector. The crucial point is an object defined by multiple parameters, where those parameters must broadcast to a common 'batch shape'.

PiperOrigin-RevId: 344312750
Change-Id: I0a98f1489a4ed24c0289f1418a7d30e81afb4c96