Making label smoothing documentation more helpful
At present, label_smoothing documentation is vague for both Categorical and BinaryCrossentropy losses.

The CategoricalCrossentropy class' documentation is currently confusing, as smoothing is implemented in a non-intuitive way. It also appears to be referring to only 2 classes (as if it were BinaryCrossentropy). The new documentation states its functionality more generally, with an example.