[TFRT:Compiler] Add the constant-op-device-assignment pass
This pass adds the device attribute to every tf.Const op based on the device attribute of the operations that read its result. If the result of a tf.Const op is read by operations placed on multiple devices, then the pass will replicate the tf.Const op once for each device.

The devices of operations are necessary to decide whether we need to insert corert.transfer op to move values between devices. And it is also useful to other compiler passes such as cluster-tf-ops-by-host, which needs to cluster operations by their host address.

PiperOrigin-RevId: 344840764
Change-Id: I64e95b2351cd0b4f019e834c4702e620985276d2