PR #45307: Extract a function for parsing operator FILL
Imported from GitHub PR https://github.com/tensorflow/tensorflow/pull/45307

Extract the parsing out of a switch statement case to create a
standalone function which can be called by the micro op resolver.

This PR is part of the work to port operator FILL from lite to micro,
as tracked in #45306.
Copybara import of the project:

--
0664214792ad2357f6224e7002661894775cb512 by Ryan Kuester <kuester@bdti.com>:

Extract a function for parsing operator FILL

Extract the parsing out of a switch statement case to create a
standalone function which can be called by the micro op resolver.

PiperOrigin-RevId: 345542606
Change-Id: I93c11fed6c9e90003b125e62fdb43479f46ea48d